+++ client/package.json
@@ -21,6 +21,7 @@
     "moment": "^2.30.1",
     "register-service-worker": "^1.7.2",
     "vue": "^3.5.16",
+    "vue-router": "^4.5.1",
     "vuedraggable": "^4.1.0"
   },
   "devDependencies": {
