+++ server/routes/cleanup.js
@@ -1,8 +1,9 @@
 import express from 'express';
 import cleanupController from '../controllers/cleanup.js';
 export const cleanupRoutes = express.Router();
+import userMiddleware from "../middleware/users.js";
 
 // POST /api/cleanup
-cleanupRoutes.post('/', cleanupController.cleanup);
+cleanupRoutes.post('/', userMiddleware.isLoggedIn, cleanupController.cleanup);
 
 export default cleanupRoutes;
